Add files via upload
There is a weird situation. The register sometimes works, but sometimes doesn't work. I guess  it is not the coding problem? If wrong coding, there will  no success on running.  It feels like the it is the probability problem? However, I did successfully registered a few users. And the login and to do tasks work very well.  Please check my video.